Refactoring: join smaller files together
