--- plugins/gatsby-source-cd-airtable/src/graphql-types/index.ts
@@ -1,2 +1,0 @@
-export * from './partner'
-export * from './project'
