--- plugins/gatsby-source-cd-airtable/src/graphql-types/partner.ts
@@ -1,7 +1,0 @@
-export const Partner = `
-  type Partner implements Node {
-    name: String!
-    url: String!
-    logoUrl: String!
-  }
-`
