--- plugins/gatsby-source-cd-airtable/src/mocks/mock-partners.ts
@@ -1,22 +1,0 @@
-import { Partner } from '../types'
-
-export const getMockPartners: () => Partner[] = () => [
-  {
-    rowId: '1',
-    name: 'Partner 1',
-    logoUrl: 'https://via.placeholder.com/400x200',
-    url: 'https://cesko.digital',
-  },
-  {
-    rowId: '2',
-    name: 'Partner 2',
-    logoUrl: 'https://via.placeholder.com/200x200',
-    url: 'https://cesko.digital',
-  },
-  {
-    rowId: '3',
-    name: 'Partner 3',
-    logoUrl: 'https://via.placeholder.com/400x200',
-    url: 'https://cesko.digital',
-  },
-]
