--- plugins/gatsby-source-cd-airtable/src/mocks/mock-projects.ts
@@ -1,55 +1,0 @@
-import { Language, Project } from '../types'
-
-export const getMockProjects: () => Project[] = () => {
-  const base = {
-    tags: ['1'],
-    contributeText: 'Lorem ipsum',
-    description: 'Lorem ipsum',
-    trelloUrl: 'https://cesko.digital',
-    githubUrl: 'https://cesko.digital',
-    url: 'https://cesko.digital',
-    coordinators: ['1'],
-    tagline: 'Lorem ipsum',
-    coverUrl: 'https://via.placeholder.com/400x200',
-    logoUrl: 'https://via.placeholder.com/100',
-  }
-
-  return [
-    {
-      name: 'První projekt',
-      rowId: '1',
-      lang: Language.Czech,
-      slug: 'prvni-projekt',
-      highlighted: true,
-      finished: false,
-      ...base,
-    },
-    {
-      name: 'First project',
-      rowId: '1',
-      lang: Language.English,
-      slug: 'first-project',
-      highlighted: true,
-      finished: false,
-      ...base,
-    },
-    {
-      name: 'Druhý projekt',
-      rowId: '2',
-      lang: Language.Czech,
-      slug: 'druhy-projekt',
-      highlighted: false,
-      finished: true,
-      ...base,
-    },
-    {
-      name: 'Second project',
-      rowId: '2',
-      lang: Language.English,
-      slug: 'second-project',
-      highlighted: false,
-      finished: true,
-      ...base,
-    },
-  ]
-}
