--- plugins/gatsby-source-cd-airtable/src/mocks/mock-volunteers.ts
@@ -1,11 +1,0 @@
-import { Volunteer } from '../types'
-
-export const getMockVolunteers: () => Volunteer[] = () => [
-  {
-    name: 'Václav Havel',
-    rowId: '1',
-    email: 'vaclav@havel.cz',
-    company: 'Czechia',
-    profilePictureUrl: 'https://via.placeholder.com/100',
-  },
-]
