+++ plugins/gatsby-source-cd-airtable/src/node-conversion.ts
